Updated imports, rename authContext to make sure expo ignores it
