+++ Studying-App/app/_layout.tsx
@@ -1,5 +1,5 @@
 import { Drawer } from "expo-router/drawer";
-import { AuthProvider, useAuth } from "./authContext";
+import { AuthProvider, useAuth } from "./components/_authContext";
 import CustomDrawer from "./components/CustomDrawer";
 
 export default function Layout() {
@@ -9,7 +9,7 @@ export default function Layout() {
         <Drawer.Screen name="index" options={{ title: "Home" }} />
         <Drawer.Screen name="dashboard" options={{ title: "Dashboard" }} />
         <Drawer.Screen name="settings" options={{ title: "Settings" }} />
-        <Drawer.Screen name="login" options={{ title: "Sign In", headerShown: true }} />
+        <Drawer.Screen name="login" options={{ title: "Sign In"}} />
       </Drawer>
     </AuthProvider>
   );
