+++ Studying-App/app/components/CustomDrawer.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from "../authContext";
+import { useAuth } from "./_authContext";
 import { View, Text, TouchableOpacity, StyleSheet } from "react-native";
 import { useRouter } from "expo-router";
