+++ Studying-App/app/login.tsx
@@ -2,12 +2,14 @@ import React, { useState } from 'react';
 import { View, Text, TextInput, Button, StyleSheet, Alert, Pressable} from 'react-native';
 import { useRouter } from 'expo-router';
 import  AsyncStorage  from "@react-native-async-storage/async-storage";
+import { useAuth } from './components/_authContext';
 
 
 const LoginScreen = () => {
  const router = useRouter()
  const [username, setUsername] = useState('');
  const [password, setPassword] = useState('');
+ const {signIn} = useAuth();
 
 
   const handleLogin = async () => {
@@ -16,6 +18,11 @@ const LoginScreen = () => {
           await AsyncStorage.setItem("username", username)
           await AsyncStorage.setItem("password", password)
           Alert.alert('Success', 'Login was successfully!');
+
+          const authToken = "token";
+          await AsyncStorage.setItem("userToken", authToken);
+          signIn(authToken);
+
           router.push({pathname: "/dashboard"});
       }
       else {
@@ -34,7 +41,7 @@ const LoginScreen = () => {
 
  return (
    <View style={styles.container}>
-     <Text style={styles.title}>Sign In</Text>
+     <Text style={styles.title}>Login</Text>
      <Text style={styles.label}>Username</Text>
      <TextInput
        style={styles.input}
